fix(GaussianState): GBS and hafnian fix
**Gaussian properties**

It is problematic, that the dimensionality (meaning scaling with HBAR)
is not trivial when calculating certain properties, therefore these
properties got refactored.

The scaling of these properties got tested.

**Gaussian Boson Sampling**

Several errors were encountered at the Gaussian Boson Sampling
calculation.
1. The Husimi covariances were not correctly calculated, this has been
   resolved by correcting the equation.
2. The calculation of the reduced A matrix was not good, since the
   subspace projection should be calculated from the subspace projection
   of the Husimi covariance, instead of calculating the A matrix on the
   whole space, and then projecting it.
3. We need to account for the non-zero displacements as well. For this,
   the `complex_displacements` property has been created, the needed
   prefactors has been correctly calculated.
4. Regarding the non-zero displacements, the loop hafnian should be
   calculated.

**Loop Hafnian**

The `loop_hafnian` function has been implemented. The logic is the same
as for the `hafnian`, but the polynomial coefficients differ. Following
this logic, the function which calculates the polynomial coefficients is
injected into the previous `hafnian` function (now renamed `_hafnian`),
and by injecting a different function, one could acquire a function for
the loop hafnian. Injection is done with `functools.partial`.