feat(gaussian): MeasureThreshold
The threshold measurement is implemented for the `gaussian` backend.

The existing logic for particle number measurement has been reused by
injecting the probability calculations in
`_apply_general_particle_number_measurement`.

Note, that it is not permitted to perform threshold detection with
non-zero displacements.

Tests, benchmarks and scripts were written to verify the soundness of
the calculation.

References:
- [Gaussian Boson Sampling using threshold detectors](https://arxiv.org/pdf/1807.01639.pdf)