fix(results): Single result for one measurement
At several measurements, for each shot a `Result` class is created. This
is redundant and wasteful.

From now on, for each measurement a single result is created.

Also, each time the term `outcome` is used, the term `sample` is used
instead, in order to avoid having multiple different names for the same
thing. The term `sample` would be more appropriate.