feat(Circuit): Support for all modes
Specifying multiple modes can be done by not specifying anything in the
`pq.Q()` instantiation. This has to be resolved at runtime, therefore
the logic of casting empty tuples into a tuple representing the whole
system is done in `Circuit.execute_instructions`.

At state registration in `pq.Q` the copying of the modes from the state
has to be supported as well. With this step, every state is required to
have a `d` attribute, which is reasonable.

After implementing this, adding support for the `all` keyword is not
hard: the `(all, )` parameter of `Q` is automatically cast to an empty
tuple.