refact(tests): Program-related tests + Expose API
The tests under `tests/api/program` got refactored to use functions
instead of classes and methods, and the remaining fixtures were moved
into a `conftest.py` file. The registration of the `Fake*` classes was
done using the `Plugin` class, therefore the registered classes got
available as an attribute of `pq`.

The intention of this is also to unify these fixtures with the ones in
`TestProgramJSONParsing`.

For writing the plugin, some parts of the API got exposed, namely
`State`, `Instruction` and `Circuit`.

Reflecting the changes, the test files got renamed.