feat(gates): Autoscaling
The autoscaling of certain gates have been implemented in this patch.

The `_ScalableBogoliubovTransformation` Instruction class has been
implemented with the single `_autoscale` method. When the instructions
are executed, this `_autoscale` method is called to rescale the
parameters of the instruction.

The parameters got rewritten with the help of `np.atleast_1d`, which
forces the variable into at least a one-dimensional array.

Some existing tests have been rewritten to use this feature.

`test_instruction_autoscaling.py` has been created to test the
autoscaling of these gates with the `fock` and `gaussian` backends.