refact(Program): Results stored in Circuit
The `results` instance attribute of `Program` is moved to `Circuit`
instead to be closer to the execution itself.

The `circuit` is made private by renaming it `_circuit`. Where there
were assertions in tests using `_circuit`, it was deleted (when
possible), since `_circuit` is a private implementation detail, and we
shouldn't care about it in tests.