refact(api.mode): Move complex logic from `Q`
There are a lot of branching and type separation in `Q.__or__`. The
different branches refer to registering certain object to the `Program`
in the `_context`.

It would be better however, if the to-be-registered objects would
define, how they would like to be registered on the `Program`. This is
better, since

a) The `Q` class is not polluted;
b) The logic of registering to the program is more relevant in the
object which we want to register;
c) By subclassing `_RegisterMixin`, anything could be registered to the
program.
d) On the existing classes, by subclassing one could achieve a different
way of registering the object easily.

Simply put, this method is more scalable.

A `copy` method is also included in `_RegisterMixin`, since most of
these objects need to be copied during the registration process to avoid
errors by changing an object by reference.