feat(fock): Squeezing and Displacement
The `Squeezing` and `Displacement` gates got implemented for all the states
in the `fock` backend.

For this, equations for the Hermite-Kampe de Fériet polynomials got
implemented under `piquasso/_math`. These polynomials can be used to
construct the operators to be applied.

Since the result of application of the operators is really hard to
verify, `validate()` methods got implemented for the distinct states.

There was a small but significant error in `is_unitary` which resulted
in false positive results, and it got fixed in this patch.

References:
- [On the role of Hermite-like polynomials in the Fock representations
  of Gaussian states](https://arxiv.org/abs/1905.10873)