refact(gates): Remove subclassing
Subclassing publically available gates from public gates does not work
well, because the parameters of the gates mix this way. E.g. at
`PositionDisplacement` it is confusing to speak about `alpha`. The
parameter turns out to be a parameter named `x`, but when subclassing
`Displacement`, the parameter gets stored as `alpha`.

To avoid this, the subclassing from public gates got removed. Private
gates are preferred to be used instead.