feat(_backends): `get_fock_probabilites`
The `get_fock_probabilities` method has been implemented for the
`GaussianState` and `*FockState` classes, which returns the particle
number detection probabilites, given a cutoff.

The motivation for doing this is to compare the `fock` and `gaussian`
backends by testing the resulting particle number detection probability.
This is the only quantity which could practically be tested.

Note, that in order to avoid probability leaking, the `Squeezing` and
`Displacement` instructions renormalize the state in the `fock` backend.
Therefore, the probabilities are only proportional in the `fock`
backend's case.

There was a small error in `block_reduce`: the specified reduction
indices are mutable, and is modified with the function call, causing
some confusion. It has been rewritten not to mutate the indices.