fix(fock): Fock space ordering
The ordering of the Fock space is not lexicographical in the particle
subspaces. This was not a problem before, since (hopefully) nothing
depended on the ordering of the Fock basis.

However, when using e.g. `fock_probabilities`, the user acquires the
result in some order. It would be better, if the order would be
lexicographic in a particle subspace.

Ordering between particle subspaces is not desired, since the ordering
would vary with different cutoffs specified.

To enable lexicographic ordering:
 - `FockBasis.create_all` got refactored to use `partitions` under
   `_math.combinatorics`;
 - Return value of `partitions` got sorted before returning.