fix: Symplectic matrix block renaming
In the code, the variables with the name
`[passive|active]_representation` actually refer to submatrices of the
general symplectic representation of a certain gate, but the
`_representation` attribute is somewhat misleading. It got renamed to
`_block` to avoid confusion.

The parameters of `GaussianTransform` (namely `P` and `A`) got renamed
to `passive` and `active` to indicate that these are blocks
corresponding to the passive or active block of the general symplectic
matrix.