fix(gate=Beamsplitter): Compatible definition
**Problem**

The current beamsplitter gate definition is not exactly compatible with
the SF definition. This wouldn't be a problem, but since we are
supporting BlackBird as well, users would expect that a `.xbb` file
results in the same output.

**Solution**

The beamsplitter gate's `passive_block` got transposed, essentially.
This results in the same gate as SF's `BSgate`. Tests were rewritten
accordingly.