fix(Circuit): Validate measured modes
Up until now, the already-measured modes could be re-introduced into the
circuit, since there was nothing to prevent it. Note, that in most
physical situations these modes cannot be reused.

The issue is fixed by maintaining a set of measured modes in the
`Circuit` class, and validating each instruction as we execute the
program.