fix(api.errors): `PiquassoException` base class
It can be handy for the user to derive all custom Piquasso exceptions
from a single base class.

As an example, a problem could arise when the user would like to catch
all errors Piquasso raises with a code like:

```
try:
    program = pq.Program(...)
    results = program.execute()
except ??? as error:
   # Handle errors raised by Piquasso
   ...
```

Here, the `???` is implemented as `PiquassoException`. Of course, one
could list all the possible Exception classes under
`piquasso.api.errors`, but this is way nicer.

Also, there were some code where `RuntimeError` was raised, probably
from the time when there were no custom exceptions. These were rewritten
to use base classes of `PiquassoException` instead.

**Registry**

In `core.registry`, the raising of `NameError` is odd, since `NameError`
is typically raised when a name is not found, where the term *name*
means a variable to refer to an object.

The `_retrieve_class` function is not even needed: one could access the
`_items` `dict` from the module as well. To simplify, the function is
deleted, and all the usages of `_retrieve_class` is rewritten.

The `registry` module got renamed to `_registry` and used as a whole
module throughout Piquasso.