feat(setup.py): Setuptools instead Poetry
Poetry is great, but for public project it is mean to require the
contributor to install `poetry`. Therefore it is deleted in favour of
`setuptools`.

Every development requirement got frozen.

The `.gitlab-ci.yml` could be updated too, but instead it got deleted,
since this project doesn't use GitLab anymore.

The `README.md` file has been rewritten accordingly. This might need a
bit more information in the future.