feat(piquasso): Docstring and hints
The docstring of an object is defined by its `__doc__` class attribute,
and attribute hints are defined in its `__dir__` method.

The `Piquasso` module wrapper class is used to wrap the
`piquasso` module to override the module's `__getattr__` function, but
some other double underscore attributes need to be redefined, like
`__doc__` and `__dir__`, which is done in this patch.

A small docstring has been written for the main module as well.