refact(*Circuit): Inject State into calculation
**Disclaimer**: This is the first commit of a considerable amount of
refactor.

**Problem**

- The calculations are too tightly coupled to the `State` classes. It
  would be great to loosen the coupling for easier integration with
  `piquassoboost`.

- In `Circuit` methods, private methods are called on a `State` object.
  This violates that single underscore methods are only used in the same
  class. By violating this, refactoring code gets harder;

- The `*State` implementations are huge and difficult to understand;

- In the `*State` classes two fundamentally different things are
  defined: calculations (corresponding to instructions, like
  `_apply_linear()`) and public methods (like `validate()`,
  `normalize()`);

- `get_instruction_map()` is odd, since it is an instance method
  returning information which is available without an instance;

- Both `Program` and `Circuit` has a reference to the actual state,
  which is somewhat confusing and redundant.

**Solution**

The current `state` instance is injected in each method of `Circuit`
corresponding to an `Instruction`.

The `get_instruction_map` is removed in favour of a class attribute
called `instruction_map`, which only refers to the names of the methods
corresponding to instructions. `execute_instructions` is rewritten
accordingly.

To further enforce the state injection, the `state` attribute of
`Circuit` has been deleted. Instead, the `state` instance is passed from
`Program.execute()` to `Circuit.execute_instructions()`.

Some `api`-related tests have been rewritten accordingly.

The `.flake8` configuration has been extended to exclude the `.venv`
folder for convenience.