feat(docs): Instructions + Boson sampling tutorial
The docstrings of some Instructions got updated.

A simple example for the boson sampling has been added.

There was some issue with the `nbsphinx` development dependency,
therefore it got updated.

The `.ipynb_checkpoints` folders got deleted and added to `.gitignore`,
since they are containing older saves of IPython notebooks which are
quite redundant in a version control system.