fix(backend=sampling): BoSS update, gate fix
There was an issue in BoSS, which got fixed in the `1.0.4` release, the
`setup.py` got updated accordingly.

There was an issue with the gate matrix application: one has to embed
the subspace of application in an identity and *then* apply it to the
whole interferometer, rather than just multiply the matrix on the
subspace. The two give different results, and the first is the good
solution.

The method `SamplingState.get_fock_probabilities()` got rewritten to
match the outputs of the other `*State.get_fock_probabilities()`
methods.

A bunch of tests have been written testing the equivalence of the
`SamplingState` and the `PureFockState`.

(+ Minor fix: `.venv` has been added to `.flake8`'s excluded
directories.)

Resolves #18