fix(gaussian): Improve numerical stability
The `ParticleNumberMeasurement` instruction had issues, since the
probability could turn out to be *slightly* less than 0.0 due to
numerical inaccuracies. The following calculations expect the
probabilities to be positive. Getting some negative probabilities could
result in errors rarely, when the `choice` turned out to be `None` due
to the fact that a non-increasing sequence of numbers are passed as
cumulative distribution to `choose_from_cumulated_probabilities`.

To resolve this issue, the positivity of probabilities are enforced in
`get_probability`.