fix(gaussian): Fulfill uncertainty relation
The Robertson-Schrödinger uncertainty relation was inaccurate in
`GaussianState.validate`, since it didn't account for the value of the
Planck constant. When one takes into account, several issues emerge:

0. `is_positive_semidefinite` strictness

Before this patch, the function `is_positive_semidefinite` did not
account for eigenvalues that are very close to 0. Therefore when a
negative value appeared among the eigenvalues e.g. in the order of
magnitude `1E-18`, the validation failed. These kind of values should
also be accepted, they technically count as `0`.

1. *Dyne measurement remaining state

The detection covariance needs to be scaled along with the value of the
Planck constant (`HBAR`) in
`GaussianState._apply_generaldyne_measurement`. If it is not scaled, the
covariance of the resulting state might not fulfill the uncertainty
relation.

When this issue got fixed, some `GeneraldyneMeasurement`-related tests
got broken, because the specified `detection_covariance` also did not
fulfill the uncertainty relation. The tests got rewritten with a valid
detection covariance, and a validation has been put in the constructor
of `GeneraldyneMeasurement` to forbid non-physical detection
covariances.

2. `TestGaussianStateOperations`

`TestGaussianStateOperations` worked with non-physical covariance. This
has been resolved by effectively rewriting the test to use actual gates.

3. Asset handling

Some test assets under the `assets` folders are slightly inaccurate.
When the uncertainty relation got corrected, these inaccuracies caused
tests failing.

This issue is resolved by increasing the precision of printing the state
when the state is saved to a file.

Resolves #15