fix(gaussian): Particle number detection errors
The `GaussianState.husimi_cov` was not correct, it did not return the
value according to the equation (1) in [Quadratic speedup for simulating
Gaussian boson sampling](https://arxiv.org/pdf/2010.15595v3.pdf).

This issue has been resolved in this patch after thorough investigation.
A test case has been written to test the defining equation, and several
tests got rewritten. `husimi_cov` got renamed to `complex_covariance`
for clarity.

Some calculations got rewritten accordingly. Notably, an odd conjugation
got removed from `calculate_particle_number_detection_probability` and
`calculate_threshold_detection_probability` at the calculation of `A`
and `OS`, respectively.

In the particle number detection case, the normalization with the
displacement got rewritten, as well as the value of `gamma`.

A test assertion has been deleted in `test_example_states.py`, since we
do not want to compare this part of the calculation to `thewalrus`: the
final results might be the same, but the partial results might differ.

The `threshold_*` scripts under `scripts` got their `shots` increased to
2000 to make the histogram and the hypothesis testing stronger.

**NOTE**

- All the covariances in `piquasso` are twice the value in Xanadu's
  articles and in `strawberryfields`, but corresponds to the definition
  in e.g. Alessio Serafini's Quantum Continuous Variables.

- In the referenced article below, it seems that sometimes `[Q]` means
  the transposed of `Q`. This seems to be the only sensible way the
  equations work out.

References:
- [Quadratic speedup for simulating Gaussian boson
  sampling](https://arxiv.org/pdf/2010.15595v3.pdf)