feat(backend=fock): Reduced states
The method `reduced` got implemented for the `*FockState` classes. The
reduction is implemented by partial tracing the density matrix in the
`FockState`.

Generally, when partial tracing a pure state, the resulting state is
mixed. Therefore, before reducing a `PureFockState`, one should convert
it to a `FockState`. The same could also be said about the
`PNCFockState`.

The `from_pure` has to be reimplemented to work with any `BaseFockState`
subclass instance. Thereby it got renamed to `from_fock_state`.