fix(gaussian): Convenient Wigner function inputs
The inputs of the Wigner function in `GaussianState` should be specified as
separate position and momentum values instead of an array of (x,p) pairs. This
method is more logical, since the wigner function calculation is mostly needed
for plotting, where a grid spanned by the position and momentum values are displayed.

With this logic, less data processing is needed in `piquasso-ui`, and less
amount of data needs to be passed, rendering the calculation faster.

Since the Wigner function can have multidimensional inputs (x, p
coordinates), a list of list has to be specified as `positions` and
`momentums` arguments.