fix(Program): `shots` is `execute` parameter
Previously, the `shots` is specified for each measurement. This wasn't
convenient, since:

- The `shots` validation logic should be duplicated for each
  measurement;

- Specifying shots for a measurement is illogical in some sense: the
  shots specifies how many time the *whole* program should be executed.

Note:

There was a small issue in `Instruction.__repr__` around the incorrect
use of `hasattr` on `Instruction.params`: the instance always has an
attribute `params`. This has also been fixed in this patch, since it
became very apparent with the increased number of parameterless
instructions.