fix(api): Generating numpy array parameter
The code generation where a parameter was a numpy array didn't work,
since the `__str__` method of `numpy.ndarray` didn't result in
executable code. The `__repr__` returns regular executable code, one
only has to prepend the generated string with `"np."`, and include
`"import numpy as np"` in the beginning of the script.

Resolves #23