feat: Program decoupled from State
The `Program.state` attribute got deleted, since it makes much more
sense if a `Program` instance is somewhat stateless. In accordance, the
`State` instance is not registered in the `Program`, instead it is
instantiated independently, then the `State.apply` method could be
called with a `Program` instance to apply a program to the state.

The instantiation from dicts got rewritten: from now on, it
automatically searches for the proper class to instantiate in the
registry. To signify this, the mixin got renamed to `_DictMixin`, and
`from_properties` got rewritten to `from_dict`. This naming is also more
comprehensible, anyway. The `properties` key in the dict also got
renamed to `attributes`, because e.g. the `property` is reserved for
something else in Python. Also, the `params` key got renamed to
`constructor_kwargs`, and the missing nesting with the constructor
kwargs in the `State`'s dict got added.

The `Program.result` property got removed, since it is not needed.

The `create_instance_from_mapping` function is removed, since it is
coupled to the `_DictMixin` by `from_dict`. Where this functionality is
needed, it is explicitely coded.

`Program.from_json` got removed, since it is not currently used in
`piquasso-ui`.

In the code generation, the state simply got removed.

All the tests concerned with this change got rewritten, and some of them
got deleted, since they got irrelevant.

Resolves #56