feat(decompositions): Williamson
The Williamson decomposition is implemented in this patch.

`takagi.py` got renamed to `decompositions.py`, and `test_takagi.py` got
renamed to `test_decompositions.py`.

The `xp_symplectic_form` is defined for verifying that the
Williamson decomposition truly returns a symplectic matrix in the xp
basis. Moreover, the `is_symplectic` got generalized for usage with
multiple forms, depending on the basis.

The symplectic group-related stuff got moved to `piquasso._math.symplectic`.