Adding type hints
The project has been refactored so that it uses type signatures in every function in the `piquasso` module.
(Note that tests are not annotated.)

The return type of `__init__` is annotated with `-> None` based on:
https://www.python.org/dev/peps/pep-0484/#the-meaning-of-annotations

From numpy>=1.21 it is possible to annotate the `dtype` of the given `np.ndarray` as `np.typing.NDArray[np.type]`.
However, strawberryfields requires numba, which does not yet support numpy>=1.21, so it is not implemented in such style but can be later when numba will support numpy==1.21.

When the type annotation refers to its class, we use the name of the class because it is only supported from 3.9 to use the class itself.

There are some instances of using:
```py
import typing

if typing.TYPE_CHECKING:
    import ...
```
This is requires since using type annotations can create circular imports but since the `TYPE_CHECKING` constant is always `False` at runtime the import won't be evaluated, but mypy (and other type-checking tools) will evaluate the contents of that block.
(https://stackoverflow.com/questions/39740632/python-type-hinting-without-cyclic-imports)

A `.mypy.ini` file has been created.

Executing `mypy` has been added to `tox.ini`.

Closes #55.