feat(GaussianState): GBS improved
The GBS or particle number detection got improved by Williamson
decomposition and heterodyne sampling. The required functionalities
(sampling, evolution) got moved from the method responsible for the
`GeneraldyneMeasurement`, and used to write the new algorithm under
`GaussianState._apply_particle_number_measurement`.

For the algorithms, setters were written for the `xxpp_mean_vector` and
`xxpp_covariance_matrix`.

For properly seeding the simulator, a `numpy` random number generator
object is stored in `constants`, which is used to generate random values
from the multivariate normal distribution. Note, that the random number
generation turned out to be not machine-independent, therefore the test
testing the results of a GBS simulation with a particular seed is
deleted.

A new script for testing the distribution equality with
`strawberryfields` in the one-mode scenario is added to `scripts`.

References:
- [Quadratic speedup for simulating Gaussian boson
  sampling](https://arxiv.org/pdf/2010.15595.pdf)
- [The Boundary for Quantum Advantage in Gaussian Boson
  Sampling](https://arxiv.org/pdf/2108.01622.pdf)