feat(GaussianState): Threshold using hafnian
The `ThresholdMeasurement` can be performed via the GBS algorithm using
the hafnian, too. This is beneficial, since the latter is faster
asymptotically than the algorithm with the torontonian. The algorithm
works by mapping the non-zero results from the GBS samples to 1.

Both calculations are enabled by the `use_torontonian` flag in
`constants`, defaulted to `False`. This is not specified in the
constructor parameter of `ThresholdMeasurement`, since it is specific to
the simulation, but (should be) irrelevant regarding the results.

The `cutoff` optional parameter is added to `ThresholdMeasurement`.

**Notes**:

- The `cutoff` parameter is odd in `ThresholdMeasurement`, while the
  `use_torontonian` is technically a global variable. We might want to
  rethink this.

- The simulation using the hafnian doesn't seem to give exactly the same
  results as with the torontonian, but the results seem to be improving
  by increasing the cutoff.