feat(api): `Simulator`
We want to separate the concerns in piquasso by the following:

- `Program` should only know about the instructions;
- `State` should only know about the current state of the quantum
  system;
- And something should define how the instructions are executed, called
  the `Simulator`.

The idea behind this is that (hopefully) if we'll use piquasso with a
real quantum computer, it would be confusing to instantiate a `State`
class: we know nothing about the actual state. The thing which executes
the instructions on a quantum computer should be some kind of client,
instead. This motivates the `Computer` class.

Moving along the same thought, one could subclass `Simulator` classes
which imitate an actual quantum computer. An abstraction layer is put
between `Program` and `State`, namely the subclasses of `Computer`.

The `State` at the end of a simulation is returned within the `Result`
class, along with the generated samples (if there are any).

It is important, that the `Config` class should be copied in the
constructor of the `State` and `Simulator` due to its mutability. When
the state is simulated with a certain configuration, the same
configuration should be used when config-dependent methods are called on
the state.

Therefore, the program execution
```
state = pq.GaussianState(d=3, config=pq.Config(hbar=1))
state.apply(program, shots=1000)
```
is changed to
```
simulator = pq.GaussianSimulator(d=3, config=pq.Config(hbar=1))
result = simulator.execute(program, shots=1000)
```

Almost all the tests got refactored accordingly.