feat(api): Simulation constants moved to Config
**Problems**

- The program definition is dependent on `constants` via
  `PositionDisplacement` and `MomentumDisplacement`. I.e. simulations
  with different HBAR should result in effectively the same state given
  a specific program containing one of the above instructions.
- Setting the simulation constants can become messy when it is done
  several times in a script, it is harder to avoid interference.
- Can have negative consequences when the constant is not reset to its
  default value.

**Solution**

The simulation constants are global variables in `constants`. These
would be much more manageable if they would be bound to the object which
performs the simulation. The constants can be put into a `Config`
dataclass containing all the necessary variables with their defaults.

The `Config` class is instantiated with their default values in
`State.__init__`, but can be reassigned easily if needed.

The gates `PositionDisplacement` and `MomentumDisplacement` can be made
HBAR-independent, when we explicitely state that the specified
parameters will be rescaled during the simulation according to the
`config` defined in the `*State`.