feat(State): `apply_instructions`
**Problem**

The `BaseFockState.from_number_preparations` is not very specific, it
could be generalized to any state. Moreover, it is quite disturbing to
have another constructor for the `BaseFockState` subclasses.

**Solution**

The `State.apply_instructions` method has been created, which has been
refactored from `State.apply`.

Moreover, in `apply_instructions`, if the instruction doesn't have a
`modes` attribute, then it is assumed that the instruction is applied
to all the modes.