feat: Delete module override of main module
**Problem**

A huge problem with `piquasso` is that the main module is overridden
with a custom class in order to override `__getattr__` to access items
from the `core._registry`. This enables the user to override classes in
`piquasso` dynamically.

Dynamic overriding is not necessarily needed: the idea behind this was
that in a script, one should be able change the underlying algorithms
with a `pq.use(SomePlugin)` call. Without this, one would have to change
the underlying classes (typically `State` subclasses). However, this can
easily be parametrized in the script as well.

Moreover, it can also cause major confusion: at first sight it is not
that apparent, that a different implementation is running than the
original.

**Solution**

Overriding the main module has been deleted. The `registry` is kept,
since we need to resolve classes from dicts and blackbird code, still.

We also want to support overriding classes, since it is also desired to
JSONs from an API call in the future, and it is convenient not to change
the JSON schema in the UI with the change of classes in the backend.