feat: Delete `Plugin` and `_registry`
**Problem 1**

A huge problem with `piquasso` is that the main module is overridden
with a custom class in order to override `__getattr__` to access items
from the `core._registry`. This enables the user to override classes in
`piquasso` dynamically.

Dynamic overriding is not necessarily needed: the idea behind this was
that in a script, one should be able change the underlying algorithms
with a `pq.use(SomePlugin)` call. Without this, one would have to change
the underlying classes (typically `State` subclasses). However, this can
easily be parametrized in the script as well.

Moreover, it can also cause major confusion: at first sight it is not
that apparent, that a different implementation is running than the
original.

**Problem 2**

Since the `State` classes has been removed from the program definition
recently, there is apparently no reason to register any `State` class in
the registry, as it is not required to parse it from a JSON when loading
a program.

One could point out that the `State` classes could be instantiated from
a JSON as well, but it is kind of useless too, since all `State`
subclasses have the same constructor signature.

Moreover, it is also not required to generate code from a `State`
subclass, since - again - it is not part of the program definition.

**Solution**

Overriding the main module has been deleted, along with the `Plugin`
class.

Since only the `Instruction` subclasses are needed to be registered,
`_registry` got deleted. Instead, the subclasses of `Instruction` is
collected with the `Instruction.__init_subclass__` hook. The subclasses
collected with this hook can be used e.g. at program instantiation from
a JSON file.

Also, the `DictMixin` and `CodeMixin` got deleted from the parent
classes of `State` along with the implemented methods, since they are no
longer needed.

The unit tests got rewritten accordingly.