fix(preparations): Supporting all iterables
**Problem**

If lists are specified as `bra` or `ket` constructor parameters in
`DensityMatrix`, it will crash, since we expect the parameters to be
tuples during the simulation.

**Solution**

The iterable is cast to a tuple explicitely. It is done for
`StateVector` as well to be sure.

The constructor signature of `DensityMatrix` has been fixed.

Some tests were rewritten, and a new test case has been added which
parses a dict with `DensityMatrix` in it.