feat(Simulator): Callables in instruction map
**Problem**

The calculations corresponding to instructions are listed in
`Simulator._instruction_map`. If one is needed to override a certain
calculation, one needs to subclass a `State`. However, this is error
prone, since the name of the registered method could easily change, and
then the original (parent class) implementation runs instead of the
desired one.

On the other hand, it is odd that the `State` classes contain the
definition on how the instructions should be executed. The methods in
these classes can be divided disjointly into property-related methods
(`xpxp_covariance_matrix`, `get_particle_detection_probability` etc.)
and the before-mentioned instruction-related methods.

**Solution**

The backends got refactored so that the calculations are explicitely
registered in the `_instruction_map` as callables. These callables are
put into their separate `calculations.py` files.