feat(Simulator): Calculations return `Result`
Having `samples` as the attribute of the `State` class is odd, since the
two should have nothing to do with each other.

More generally, the `State` classes should not have any memory regarding
the simulation whatsoever, which makes sense from the perspective of a
user as well.

Therefore the `State.samples` attribute got deleted, and the samples are
optionally returned by the calculations defined in the `Simulator`
classes. The signature of the calculation functions got rewritten to
return with a `Result` instance.

The `Simulator.execute_instructions` got rewritten to return with a
`Result` instead of a `State`, since when a measurement is also
specified, the samples would only be accessible by `State.samples`,
which is deleted in this patch.