fix(fock): Explicit casting of complex coefficients
During the calculation of the `ParticleNumberMeasurement` instruction in
the `fock` backends, the elements of the state vector or density matrix
is cast with `float`. This wouldn't be a problem where all the modes are
measured, but it definitely doesn't work for less modes in general,
because (essentially) the imaginary parts are thrown away.

The algorithm has been rewritten to reduce the state to the measured
modes first, and then calculate the particle number measurement
probabilities corresponding to the reduced state.