fix(gaussian): Config inheritance of states
There were 3 instances in the Gaussian backend, where the config is not
properly passed to a new instance of the state. This caused unusual
validation errors due to covariance matrices not adhering to the
Robertson-Schrödinger uncertainty relation, which turned out to be an
issue because the covariance matrices were divided with the default
`hbar` value, instead of the one in the config.

This has been fixed in this patch by passing the config to every
instantiation.

Several test cases were added, which would have failed with the original
code.