perf(fock/pure): Avoid embedding of passive gates
Some gate implementation embedded the matrix into the larger Fock space,
then multiplied the state vector, which is quite inefficient. It is
already been rewritten for the active gates in commit
0f215746cf75eb0252120257792e851536eeb514, the passive gates are handled
in this patch.

Note, that `_apply_single_mode_representation_to_state` is rewritten to
work with multiple modes, hence the name
`_apply_subsystem_representation_to_state`.

This will be done for the `general` Fock simulation later.