feat(sampling): TransmissivityMatrix
The channel `TransmissivityMatrix` might be handy if the user wants to
model more general losses in the `SamplingSimulator`. With this one can
specify a general transmissivity matrix, which is essentially equivalent
to a `Loss` channel with two (generally different) `Interferometer`
gates before and after.

Tests have been written. For these, and `__eq__` method is constructed
in `SamplingState`.