feat: Reversing occupation number order
The aim of this patch is to reverse the order of the occupation numbers
in the fixed particle number subspaces, because we want to order e.g.
the one-particle occupation numbers as
```
| 100 >, | 010 >, | 001 >;
```
instead of the opposite order, as it is now.

There are two places where the order of the occupation numbers are
defined:
1. The `partitions` function under `_math`;
2. `theboss` dependency.

1. is modified to reverse the order. The dependency `theboss` has also
been modified to follow this order in version 2, so this version is
updated in `setup.py`.

The tests were rewritten accordingly.